BPMemory: Cleanup GenMode and fix incorrect bitfield size.

+++ Source/Core/VideoCommon/BPStructs.cpp
@@ -91,9 +91,9 @@ void BPWritten(const BPCmd& bp)
 	case BPMEM_GENMODE: // Set the Generation Mode
 		{
 			PRIM_LOG("genmode: texgen=%d, col=%d, multisampling=%d, tev=%d, cullmode=%d, ind=%d, zfeeze=%d",
-			         bpmem.genMode.numtexgens, bpmem.genMode.numcolchans,
-			         bpmem.genMode.multisampling, bpmem.genMode.numtevstages+1, bpmem.genMode.cullmode,
-			         bpmem.genMode.numindstages, bpmem.genMode.zfreeze);
+			         (u32)bpmem.genMode.numtexgens, (u32)bpmem.genMode.numcolchans,
+			         (u32)bpmem.genMode.multisampling, (u32)bpmem.genMode.numtevstages+1, (u32)bpmem.genMode.cullmode,
+			         (u32)bpmem.genMode.numindstages, (u32)bpmem.genMode.zfreeze);
 
 			// Only call SetGenerationMode when cull mode changes.
 			if (bp.changes & 0xC000)
